Move incentive methods from Ads to dedicated AdsIncentives class
